+++ projects/generic/routes/test/v1/testRoutes.js
@@ -6,7 +6,7 @@ module.exports = {
     name: 'Tests',
     version: ver,
     prefix: '/api/tests/generic/v1',
-    environment: ["local","qa"],
+    environment: ["local","int"],
     routes: {
 
         //users